Refactor scene and asteroid modules
