+++ src/index.html
@@ -35,6 +35,6 @@
             <p id="planetDetails"></p>
         </div>
     </div>
-    <script type="module" src="./script.js"></script>
+    <script type="module" src="./main.js"></script>
 </body>
 </html>
