--- src/script.js
@@ -1,1204 +1,0 @@
-import * as THREE from 'three';
-import * as dat from 'dat.gui';
-import { OrbitControls } from 'three/examples/jsm/controls/OrbitControls.js';
-import { EffectComposer } from 'three/addons/postprocessing/EffectComposer.js';
-import { UnrealBloomPass } from 'three/addons/postprocessing/UnrealBloomPass.js';
-import { RenderPass } from 'three/addons/postprocessing/RenderPass.js';
-import { GLTFLoader } from 'three/examples/jsm/loaders/GLTFLoader.js';
-import { OutlinePass } from 'three/addons/postprocessing/OutlinePass.js';
-
-import bgTexture1 from '/images/1.jpg';
-import bgTexture2 from '/images/2.jpg';
-import bgTexture3 from '/images/3.jpg';
-import bgTexture4 from '/images/4.jpg';
-import sunTexture from '/images/sun.jpg';
-import mercuryTexture from '/images/mercurymap.jpg';
-import mercuryBump from '/images/mercurybump.jpg';
-import venusTexture from '/images/venusmap.jpg';
-import venusBump from '/images/venusmap.jpg';
-import venusAtmosphere from '/images/venus_atmosphere.jpg';
-import earthTexture from '/images/earth_daymap.jpg';
-import earthNightTexture from '/images/earth_nightmap.jpg';
-import earthAtmosphere from '/images/earth_atmosphere.jpg';
-import earthMoonTexture from '/images/moonmap.jpg';
-import earthMoonBump from '/images/moonbump.jpg';
-import marsTexture from '/images/marsmap.jpg';
-import marsBump from '/images/marsbump.jpg';
-import jupiterTexture from '/images/jupiter.jpg';
-import ioTexture from '/images/jupiterIo.jpg';
-import europaTexture from '/images/jupiterEuropa.jpg';
-import ganymedeTexture from '/images/jupiterGanymede.jpg';
-import callistoTexture from '/images/jupiterCallisto.jpg';
-import saturnTexture from '/images/saturnmap.jpg';
-import satRingTexture from '/images/saturn_ring.png';
-import uranusTexture from '/images/uranus.jpg';
-import uraRingTexture from '/images/uranus_ring.png';
-import neptuneTexture from '/images/neptune.jpg';
-import plutoTexture from '/images/plutomap.jpg';
-
-// ******  SETUP  ******
-console.log("Create the scene");
-const scene = new THREE.Scene();
-
-console.log("Create a perspective projection camera");
-var camera = new THREE.PerspectiveCamera( 45, window.innerWidth/window.innerHeight, 0.1, 1000 );
-camera.position.set(-175, 115, 5);
-
-console.log("Create the renderer");
-const renderer = new THREE.WebGL1Renderer();
-renderer.setSize(window.innerWidth, window.innerHeight);
-document.body.appendChild(renderer.domElement);
-renderer.toneMapping = THREE.ACESFilmicToneMapping;
-
-console.log("Create an orbit control");
-const controls = new OrbitControls(camera, renderer.domElement);
-controls.enableDamping = true;
-controls.dampingFactor = 0.75;
-controls.screenSpacePanning = false;
-
-const earthDefaultCameraOffset = new THREE.Vector3(-90, 45, 140);
-const earthDefaultCameraPosition = new THREE.Vector3();
-const earthDefaultTargetPosition = new THREE.Vector3();
-let zoomOutTargetPosition = new THREE.Vector3();
-
-console.log("Set up texture loader");
-const cubeTextureLoader = new THREE.CubeTextureLoader();
-const loadTexture = new THREE.TextureLoader();
-
-// ******  POSTPROCESSING setup ******
-const composer = new EffectComposer(renderer);
-composer.addPass(new RenderPass(scene, camera));
-
-// ******  OUTLINE PASS  ******
-const outlinePass = new OutlinePass(new THREE.Vector2(window.innerWidth, window.innerHeight), scene, camera);
-outlinePass.edgeStrength = 3;
-outlinePass.edgeGlow = 1;
-outlinePass.visibleEdgeColor.set(0xffffff);
-outlinePass.hiddenEdgeColor.set(0x190a05);
-composer.addPass(outlinePass);
-
-// ******  BLOOM PASS  ******
-const bloomPass = new UnrealBloomPass(new THREE.Vector2(window.innerWidth, window.innerHeight), 1, 0.4, 0.85);
-bloomPass.threshold = 1;
-bloomPass.radius = 0.9;
-composer.addPass(bloomPass);
-
-// ****** AMBIENT LIGHT ******
-console.log("Add the ambient light");
-var lightAmbient = new THREE.AmbientLight(0x222222, 6); 
-scene.add(lightAmbient);
-
-// ******  Star background  ******
-scene.background = cubeTextureLoader.load([
-
-  bgTexture3,
-  bgTexture1,
-  bgTexture2,
-  bgTexture2,
-  bgTexture4,
-  bgTexture2
-]);
-
-// ******  CONTROLS  ******
-const gui = new dat.GUI({ autoPlace: false });
-const customContainer = document.getElementById('gui-container');
-customContainer.appendChild(gui.domElement);
-
-// ****** SETTINGS FOR INTERACTIVE CONTROLS  ******
-const settings = {
-  accelerationOrbit: 1,
-  acceleration: 1,
-  sunIntensity: 1.9
-};
-
-gui.add(settings, 'accelerationOrbit', 0, 10).onChange(value => {
-});
-gui.add(settings, 'acceleration', 0, 10).onChange(value => {
-});
-gui.add(settings, 'sunIntensity', 1, 10).onChange(value => {
-  sunMat.emissiveIntensity = value;
-});
-
-// mouse movement
-const raycaster = new THREE.Raycaster();
-const mouse = new THREE.Vector2();
-
-function onMouseMove(event) {
-    const isCanvasInteraction = event.target === renderer.domElement;
-    if (!isCanvasInteraction) {
-        setHoveredAsteroidEntry(null);
-        return;
-    }
-
-    event.preventDefault();
-    mouse.x = (event.clientX / window.innerWidth) * 2 - 1;
-    mouse.y = - (event.clientY / window.innerHeight) * 2 + 1;
-    updateHoveredAsteroid();
-}
-
-// ******  SELECT PLANET  ******
-let selectedPlanet = null;
-let isMovingTowardsPlanet = false;
-let targetCameraPosition = new THREE.Vector3();
-let offset;
-
-function onDocumentMouseDown(event) {
-  if (event.target !== renderer.domElement) {
-    return;
-  }
-
-  event.preventDefault();
-
-  mouse.x = (event.clientX / window.innerWidth) * 2 - 1;
-  mouse.y = - (event.clientY / window.innerHeight) * 2 + 1;
-
-  updateHoveredAsteroid();
-
-  if (hoveredAsteroidEntry) {
-    handleAsteroidSelection(hoveredAsteroidEntry.data.id);
-  } else {
-    clearAsteroidSelection();
-  }
-}
-
-function identifyPlanet(clickedObject) {
-  // Logic to identify which planet was clicked based on the clicked object, different offset for camera distance
-        if (clickedObject.material === mercury.planet.material) {
-          offset = 10;
-          return mercury;
-        } else if (clickedObject.material === venus.Atmosphere.material) {
-          offset = 25;
-          return venus;
-        } else if (clickedObject.material === earth.Atmosphere.material) {
-          offset = 25;
-          return earth;
-        } else if (clickedObject.material === mars.planet.material) {
-          offset = 15;
-          return mars;
-        } else if (clickedObject.material === jupiter.planet.material) {
-          offset = 50;
-          return jupiter;
-        } else if (clickedObject.material === saturn.planet.material) {
-          offset = 50;
-          return saturn;
-        } else if (clickedObject.material === uranus.planet.material) {
-          offset = 25;
-          return uranus;
-        } else if (clickedObject.material === neptune.planet.material) {
-          offset = 20;
-          return neptune;
-        } else if (clickedObject.material === pluto.planet.material) {
-          offset = 10;
-          return pluto;
-        } 
-
-  return null;
-}
-
-// ******  SHOW PLANET INFO AFTER SELECTION  ******
-function showPlanetInfo(planet) {
-  var info = document.getElementById('planetInfo');
-  var name = document.getElementById('planetName');
-  var details = document.getElementById('planetDetails');
-
-  name.innerText = planet;
-  details.innerText = `Radius: ${planetData[planet].radius}\nTilt: ${planetData[planet].tilt}\nRotation: ${planetData[planet].rotation}\nOrbit: ${planetData[planet].orbit}\nDistance: ${planetData[planet].distance}\nMoons: ${planetData[planet].moons}\nInfo: ${planetData[planet].info}`;
-
-  info.style.display = 'block';
-}
-let isZoomingOut = false;
-// close 'x' button function
-function updateEarthDefaultView(applyToCamera = false) {
-  if (typeof earth !== 'undefined' && earth.planet) {
-    earth.planet.getWorldPosition(earthDefaultTargetPosition);
-  } else {
-    earthDefaultTargetPosition.set(0, 0, 0);
-  }
-
-  earthDefaultCameraPosition.copy(earthDefaultTargetPosition).add(earthDefaultCameraOffset);
-  zoomOutTargetPosition.copy(earthDefaultCameraPosition);
-
-  if (applyToCamera) {
-    controls.target.copy(earthDefaultTargetPosition);
-    camera.position.copy(earthDefaultCameraPosition);
-  }
-}
-
-function closeInfo() {
-  var info = document.getElementById('planetInfo');
-  info.style.display = 'none';
-  settings.accelerationOrbit = 1;
-  updateEarthDefaultView();
-  controls.target.copy(earthDefaultTargetPosition);
-  isZoomingOut = true;
-}
-window.closeInfo = closeInfo;
-// close info when clicking another planet
-function closeInfoNoZoomOut() {
-  var info = document.getElementById('planetInfo');
-  info.style.display = 'none';
-  settings.accelerationOrbit = 1;
-}
-// ******  SUN  ******
-let sunMat;
-
-const sunSize = 697/40; // 40 times smaller scale than earth
-const sunGeom = new THREE.SphereGeometry(sunSize, 32, 20);
-sunMat = new THREE.MeshStandardMaterial({
-  emissive: 0xFFF88F,
-  emissiveMap: loadTexture.load(sunTexture),
-  emissiveIntensity: settings.sunIntensity
-});
-const sun = new THREE.Mesh(sunGeom, sunMat);
-scene.add(sun);
-
-//point light in the sun
-const pointLight = new THREE.PointLight(0xFDFFD3 , 1200, 400, 1.4);
-scene.add(pointLight);
-
-
-// ******  PLANET CREATION FUNCTION  ******
-function createPlanet(planetName, size, position, tilt, texture, bump, ring, atmosphere, moons){
-
-  let material;
-  if (texture instanceof THREE.Material){
-    material = texture;
-  } 
-  else if(bump){
-    material = new THREE.MeshPhongMaterial({
-    map: loadTexture.load(texture),
-    bumpMap: loadTexture.load(bump),
-    bumpScale: 0.7
-    });
-  }
-  else {
-    material = new THREE.MeshPhongMaterial({
-    map: loadTexture.load(texture)
-    });
-  } 
-
-  const name = planetName;
-  const geometry = new THREE.SphereGeometry(size, 32, 20);
-  const planet = new THREE.Mesh(geometry, material);
-  const planet3d = new THREE.Object3D;
-  const planetSystem = new THREE.Group();
-  planetSystem.add(planet);
-  let Atmosphere;
-  let Ring;
-  planet.position.x = position;
-  planet.rotation.z = tilt * Math.PI / 180;
-
-  // add orbit path
-  const orbitPath = new THREE.EllipseCurve(
-    0, 0,            // ax, aY
-    position, position, // xRadius, yRadius
-    0, 2 * Math.PI,   // aStartAngle, aEndAngle
-    false,            // aClockwise
-    0                 // aRotation
-);
-
-  const pathPoints = orbitPath.getPoints(100);
-  const orbitGeometry = new THREE.BufferGeometry().setFromPoints(pathPoints);
-  const orbitMaterial = new THREE.LineBasicMaterial({ color: 0xFFFFFF, transparent: true, opacity: 0.03 });
-  const orbit = new THREE.LineLoop(orbitGeometry, orbitMaterial);
-  orbit.rotation.x = Math.PI / 2;
-  planetSystem.add(orbit);
-
-  //add ring
-  if(ring)
-  {
-    const RingGeo = new THREE.RingGeometry(ring.innerRadius, ring.outerRadius,30);
-    const RingMat = new THREE.MeshStandardMaterial({
-      map: loadTexture.load(ring.texture),
-      side: THREE.DoubleSide
-    });
-    Ring = new THREE.Mesh(RingGeo, RingMat);
-    planetSystem.add(Ring);
-    Ring.position.x = position;
-    Ring.rotation.x = -0.5 *Math.PI;
-    Ring.rotation.y = -tilt * Math.PI / 180;
-  }
-  
-  //add atmosphere
-  if(atmosphere){
-    const atmosphereGeom = new THREE.SphereGeometry(size+0.1, 32, 20);
-    const atmosphereMaterial = new THREE.MeshPhongMaterial({
-      map:loadTexture.load(atmosphere),
-      transparent: true,
-      opacity: 0.4,
-      depthTest: true,
-      depthWrite: false
-    })
-    Atmosphere = new THREE.Mesh(atmosphereGeom, atmosphereMaterial)
-    
-    Atmosphere.rotation.z = 0.41;
-    planet.add(Atmosphere);
-  }
-
-  //add moons
-  if(moons){
-    moons.forEach(moon => {
-      let moonMaterial;
-      
-      if(moon.bump){
-        moonMaterial = new THREE.MeshStandardMaterial({
-          map: loadTexture.load(moon.texture),
-          bumpMap: loadTexture.load(moon.bump),
-          bumpScale: 0.5
-        });
-      } else{
-        moonMaterial = new THREE.MeshStandardMaterial({
-          map: loadTexture.load(moon.texture)
-        });
-      }
-      const moonGeometry = new THREE.SphereGeometry(moon.size, 32, 20);
-      const moonMesh = new THREE.Mesh(moonGeometry, moonMaterial);
-      const moonOrbitDistance = size * 1.5;
-      moonMesh.position.set(moonOrbitDistance, 0, 0);
-      planetSystem.add(moonMesh);
-      moon.mesh = moonMesh;
-    });
-  }
-  //add planet system to planet3d object and to the scene
-  planet3d.add(planetSystem);
-  scene.add(planet3d);
-  return {name, planet, planet3d, Atmosphere, moons, planetSystem, Ring};
-}
-
-
-// ******  LOADING OBJECTS METHOD  ******
-function loadObject(path, position, scale, callback) {
-  const loader = new GLTFLoader();
-
-  loader.load(path, function (gltf) {
-      const obj = gltf.scene;
-      obj.position.set(position, 0, 0);
-      obj.scale.set(scale, scale, scale);
-      scene.add(obj);
-      if (callback) {
-        callback(obj);
-      }
-  }, undefined, function (error) {
-      console.error('An error happened', error);
-  });
-}
-
-// ******  ASTEROIDS  ******
-const asteroidListElement = document.getElementById('asteroidList');
-const asteroidPanelElement = document.getElementById('asteroidListPanel');
-const impactLegendElement = document.getElementById('impactLegendOverlay');
-
-if (asteroidPanelElement) {
-  asteroidPanelElement.dataset.component = 'asteroid-panel';
-}
-
-if (impactLegendElement) {
-  impactLegendElement.dataset.component = 'impact-legend';
-}
-
-const asteroidYieldColors = {
-  low: 0x58d68d,
-  medium: 0xf4d03f,
-  high: 0xe74c3c
-};
-
-const asteroidYieldRadiusScale = {
-  low: 1.3,
-  medium: 1.8,
-  high: 2.6
-};
-
-const asteroidDefaultCameraOffset = new THREE.Vector3(25, 15, 25);
-const asteroidFocusPoint = new THREE.Vector3();
-const asteroidCameraTarget = new THREE.Vector3();
-const asteroidWorkVector = new THREE.Vector3();
-const asteroidMidPoint = new THREE.Vector3();
-const earthWorldPosition = new THREE.Vector3();
-
-const asteroidCatalog = [
-  {
-    id: 'apophis',
-    name: '99942 Apophis',
-    designation: 'Apophis',
-    tntYieldMt: 800,
-    orbit: { semiMajorAxis: 150, eccentricity: 0.12, inclination: 6, angularVelocity: 0.00012 },
-    spinRate: 0.0015,
-    visualScale: 0.018,
-    description: 'Potentially hazardous Aten asteroid with a close 2029 approach.'
-  },
-  {
-    id: 'bennu',
-    name: '101955 Bennu',
-    designation: 'Bennu',
-    tntYieldMt: 4.5,
-    orbit: { semiMajorAxis: 120, eccentricity: 0.21, inclination: 6, angularVelocity: 0.00009 },
-    spinRate: 0.0018,
-    visualScale: 0.016,
-    description: 'Carbonaceous near-Earth asteroid sampled by OSIRIS-REx.'
-  },
-  {
-    id: 'didymos',
-    name: '65803 Didymos',
-    designation: 'Didymos',
-    tntYieldMt: 15.0,
-    orbit: { semiMajorAxis: 180, eccentricity: 0.08, inclination: 3, angularVelocity: 0.00007 },
-    spinRate: 0.0012,
-    visualScale: 0.02,
-    description: 'Binary system primary targeted by the DART mission.'
-  }
-];
-
-const asteroidEntries = asteroidCatalog.map((data, index) => ({
-  data,
-  mesh: null,
-  listItem: null,
-  orbitAngle: data.initialPhase ?? index * 1.3,
-  orbitAngularVelocity: data.orbit.angularVelocity
-}));
-
-let selectedAsteroidEntry = null;
-let hoveredAsteroidEntry = null;
-let asteroidTrajectoryLine = null;
-let asteroidImpactOverlay = null;
-let isMovingTowardsAsteroid = false;
-
-buildAsteroidList();
-initializeAsteroids();
-
-function buildAsteroidList() {
-  if (!asteroidListElement) {
-    return;
-  }
-
-  asteroidListElement.innerHTML = '';
-
-  asteroidEntries.forEach(entry => {
-    const { data } = entry;
-    const item = document.createElement('li');
-    item.className = 'asteroid-panel__item';
-    item.dataset.asteroidId = data.id;
-    item.innerHTML = `<span class='asteroid-panel__name'>${data.name}</span>` +
-      `<span class='asteroid-panel__meta'>Yield: ${data.tntYieldMt} Mt | a=${data.orbit.semiMajorAxis} | i=${data.orbit.inclination} deg</span>`;
-    item.addEventListener('click', () => handleAsteroidSelection(data.id));
-    asteroidListElement.appendChild(item);
-    entry.listItem = item;
-  });
-}
-
-function initializeAsteroids() {
-  const loader = new GLTFLoader();
-
-  loader.load('/asteroids/asteroidPack.glb', gltf => {
-    const templates = [];
-
-    gltf.scene.traverse(child => {
-      if (child.isMesh) {
-        templates.push(child);
-      }
-    });
-
-    if (templates.length === 0) {
-      createFallbackAsteroids();
-      return;
-    }
-
-    asteroidEntries.forEach((entry, index) => {
-      const source = templates[index % templates.length].clone(true);
-      if (source.geometry) {
-        source.geometry = source.geometry.clone();
-      }
-      if (Array.isArray(source.material)) {
-        source.material = source.material.map(material => material.clone());
-      } else if (source.material) {
-        source.material = source.material.clone();
-      }
-      source.scale.setScalar(entry.data.visualScale);
-      source.castShadow = true;
-      source.receiveShadow = true;
-      source.name = entry.data.name;
-
-      source.traverse(node => {
-        if (!node.userData) {
-          node.userData = {};
-        }
-        node.userData.asteroidId = entry.data.id;
-        node.userData.metadata = entry.data;
-      });
-
-      scene.add(source);
-      entry.mesh = source;
-      updateAsteroidTransform(entry, false);
-    });
-  }, undefined, error => {
-    console.warn('Failed to load asteroid GLB. Falling back to procedural meshes.', error);
-    createFallbackAsteroids();
-  });
-}
-
-function createFallbackAsteroids() {
-  const baseGeometry = new THREE.IcosahedronGeometry(2, 1);
-
-  asteroidEntries.forEach(entry => {
-    const material = new THREE.MeshStandardMaterial({ color: 0xadb5bd, flatShading: true });
-    const mesh = new THREE.Mesh(baseGeometry.clone(), material);
-    mesh.scale.setScalar(entry.data.visualScale * 8);
-    mesh.castShadow = true;
-    mesh.receiveShadow = true;
-    mesh.name = entry.data.name;
-    mesh.userData = mesh.userData || {};
-    mesh.userData.asteroidId = entry.data.id;
-    mesh.userData.metadata = entry.data;
-
-    scene.add(mesh);
-    entry.mesh = mesh;
-    updateAsteroidTransform(entry, false);
-  });
-}
-
-function getAsteroidMeshes() {
-  return asteroidEntries
-    .filter(entry => entry.mesh)
-    .map(entry => entry.mesh);
-}
-
-function findAsteroidEntryFromObject(object) {
-  let current = object;
-  while (current && !current.userData?.asteroidId && current.parent) {
-    current = current.parent;
-  }
-  const asteroidId = current?.userData?.asteroidId;
-  if (!asteroidId) {
-    return null;
-  }
-  return asteroidEntries.find(entry => entry.data.id === asteroidId) || null;
-}
-
-function setHoveredAsteroidEntry(entry) {
-  if (hoveredAsteroidEntry && hoveredAsteroidEntry.listItem) {
-    hoveredAsteroidEntry.listItem.classList.remove('asteroid-panel__item--hover');
-  }
-
-  hoveredAsteroidEntry = entry;
-
-  if (hoveredAsteroidEntry && hoveredAsteroidEntry.listItem) {
-    hoveredAsteroidEntry.listItem.classList.add('asteroid-panel__item--hover');
-  }
-
-  if (renderer && renderer.domElement) {
-    renderer.domElement.style.cursor = hoveredAsteroidEntry ? 'pointer' : 'auto';
-  }
-}
-
-function updateHoveredAsteroid() {
-  const meshes = getAsteroidMeshes();
-  if (meshes.length === 0) {
-    setHoveredAsteroidEntry(null);
-    return;
-  }
-
-  raycaster.setFromCamera(mouse, camera);
-  const intersects = raycaster.intersectObjects(meshes, true);
-  if (intersects.length === 0) {
-    setHoveredAsteroidEntry(null);
-    return;
-  }
-
-  const entry = findAsteroidEntryFromObject(intersects[0].object);
-  setHoveredAsteroidEntry(entry || null);
-}
-
-function getYieldBand(tntMt) {
-  if (tntMt <= 1) {
-    return 'low';
-  }
-
-  if (tntMt <= 10) {
-    return 'medium';
-  }
-
-  return 'high';
-}
-
-function updateAsteroidTransform(entry, advance = true) {
-  if (!entry.mesh) {
-    return;
-  }
-
-  if (advance) {
-    entry.orbitAngle += entry.orbitAngularVelocity * settings.accelerationOrbit;
-  }
-
-  const { semiMajorAxis: a, eccentricity, inclination } = entry.data.orbit;
-  const b = a * Math.sqrt(1 - Math.pow(eccentricity, 2));
-  const angle = entry.orbitAngle;
-
-  const x = a * Math.cos(angle);
-  const z = b * Math.sin(angle);
-  const inclineRad = THREE.MathUtils.degToRad(inclination);
-  const y = Math.sin(angle) * Math.sin(inclineRad) * a * 0.2;
-
-  entry.mesh.position.set(x, y, z);
-
-  const spinRate = entry.data.spinRate ?? 0.001;
-  entry.mesh.rotation.y += spinRate * settings.acceleration;
-}
-
-function updateAsteroids() {
-  asteroidEntries.forEach(entry => updateAsteroidTransform(entry));
-
-  if (selectedAsteroidEntry && selectedAsteroidEntry.mesh) {
-    updateTrajectoryLine(selectedAsteroidEntry);
-  }
-}
-
-function handleAsteroidSelection(id) {
-  const entry = asteroidEntries.find(item => item.data.id === id);
-  if (!entry || !entry.mesh) {
-    return;
-  }
-
-  if (selectedAsteroidEntry && selectedAsteroidEntry.data.id === id) {
-    clearAsteroidSelection();
-    return;
-  }
-
-  clearAsteroidSelection();
-
-  selectedAsteroidEntry = entry;
-  if (entry.listItem) {
-    entry.listItem.classList.add('asteroid-panel__item--active');
-  }
-
-  focusCameraOnAsteroid(entry);
-  createTrajectoryLine(entry);
-  createImpactOverlay(entry);
-}
-
-function clearAsteroidSelection() {
-  if (selectedAsteroidEntry && selectedAsteroidEntry.listItem) {
-    selectedAsteroidEntry.listItem.classList.remove('asteroid-panel__item--active');
-  }
-
-  removeTrajectoryLine();
-  removeImpactOverlay();
-  selectedAsteroidEntry = null;
-  isMovingTowardsAsteroid = false;
-  updateEarthDefaultView();
-  controls.target.copy(earthDefaultTargetPosition);
-  isZoomingOut = true;
-}
-
-function focusCameraOnAsteroid(entry) {
-  closeInfoNoZoomOut();
-  selectedPlanet = null;
-  isMovingTowardsPlanet = false;
-  isZoomingOut = false;
-
-  if (!entry.mesh) {
-    return;
-  }
-
-  entry.mesh.getWorldPosition(asteroidWorkVector);
-  asteroidCameraTarget.copy(asteroidWorkVector).add(asteroidDefaultCameraOffset);
-  controls.target.copy(asteroidWorkVector);
-  isMovingTowardsAsteroid = true;
-}
-
-function createTrajectoryLine(entry) {
-  removeTrajectoryLine();
-
-  const points = getTrajectoryPoints(entry);
-  const geometry = new THREE.BufferGeometry().setFromPoints(points);
-  const material = new THREE.LineBasicMaterial({
-    color: 0x9ecbff,
-    transparent: true,
-    opacity: 0.8
-  });
-
-  asteroidTrajectoryLine = new THREE.Line(geometry, material);
-  asteroidTrajectoryLine.frustumCulled = false;
-  scene.add(asteroidTrajectoryLine);
-}
-
-function updateTrajectoryLine(entry) {
-  if (!asteroidTrajectoryLine) {
-    return;
-  }
-
-  const points = getTrajectoryPoints(entry);
-  asteroidTrajectoryLine.geometry.setFromPoints(points);
-}
-
-function removeTrajectoryLine() {
-  if (!asteroidTrajectoryLine) {
-    return;
-  }
-
-  scene.remove(asteroidTrajectoryLine);
-  asteroidTrajectoryLine.geometry.dispose();
-  asteroidTrajectoryLine.material.dispose();
-  asteroidTrajectoryLine = null;
-}
-
-function createImpactOverlay(entry) {
-  removeImpactOverlay();
-
-  const yieldBand = getYieldBand(entry.data.tntYieldMt);
-  const color = asteroidYieldColors[yieldBand];
-  const scaleFactor = asteroidYieldRadiusScale[yieldBand];
-  const earthRadius = (typeof earth !== 'undefined' && earth.planet && earth.planet.geometry && earth.planet.geometry.parameters)
-    ? earth.planet.geometry.parameters.radius
-    : 6.4;
-
-  const geometry = new THREE.CircleGeometry(earthRadius * scaleFactor, 48);
-  const material = new THREE.MeshBasicMaterial({
-    color,
-    transparent: true,
-    opacity: 0.25,
-    side: THREE.DoubleSide
-  });
-
-  asteroidImpactOverlay = new THREE.Mesh(geometry, material);
-  asteroidImpactOverlay.rotation.x = -Math.PI / 2;
-  asteroidImpactOverlay.position.set(0, 0.1, 0);
-
-  if (typeof earth !== 'undefined' && earth.planet) {
-    earth.planet.add(asteroidImpactOverlay);
-  } else {
-    scene.add(asteroidImpactOverlay);
-  }
-}
-
-function removeImpactOverlay() {
-  if (!asteroidImpactOverlay) {
-    return;
-  }
-
-  if (asteroidImpactOverlay.parent) {
-    asteroidImpactOverlay.parent.remove(asteroidImpactOverlay);
-  }
-
-  asteroidImpactOverlay.geometry.dispose();
-  asteroidImpactOverlay.material.dispose();
-  asteroidImpactOverlay = null;
-}
-
-function getTrajectoryPoints(entry) {
-  if (!entry.mesh) {
-    return [];
-  }
-
-  entry.mesh.getWorldPosition(asteroidWorkVector);
-  asteroidFocusPoint.copy(asteroidWorkVector);
-
-  if (typeof earth !== 'undefined' && earth.planet) {
-    earth.planet.getWorldPosition(earthWorldPosition);
-  } else {
-    earthWorldPosition.set(0, 0, 0);
-  }
-
-  asteroidMidPoint.copy(asteroidWorkVector).lerp(earthWorldPosition, 0.5);
-  asteroidMidPoint.y += 20;
-
-  const curve = new THREE.CatmullRomCurve3([
-    asteroidWorkVector.clone(),
-    asteroidMidPoint.clone(),
-    earthWorldPosition.clone()
-  ]);
-
-  return curve.getPoints(32);
-}
-// Earth day/night effect shader material
-const earthMaterial = new THREE.ShaderMaterial({
-  uniforms: {
-    dayTexture: { type: "t", value: loadTexture.load(earthTexture) },
-    nightTexture: { type: "t", value: loadTexture.load(earthNightTexture) },
-    sunPosition: { type: "v3", value: sun.position }
-  },
-  vertexShader: `
-    varying vec3 vNormal;
-    varying vec2 vUv;
-    varying vec3 vSunDirection;
-
-    uniform vec3 sunPosition;
-
-    void main() {
-      vUv = uv;
-      vec4 worldPosition = modelMatrix * vec4(position, 1.0);
-      vNormal = normalize(modelMatrix * vec4(normal, 0.0)).xyz;
-      vSunDirection = normalize(sunPosition - worldPosition.xyz);
-      gl_Position = projectionMatrix * modelViewMatrix * vec4(position, 1.0);
-    }
-  `,
-  fragmentShader: `
-    uniform sampler2D dayTexture;
-    uniform sampler2D nightTexture;
-
-    varying vec3 vNormal;
-    varying vec2 vUv;
-    varying vec3 vSunDirection;
-
-    void main() {
-      float intensity = max(dot(vNormal, vSunDirection), 0.0);
-      vec4 dayColor = texture2D(dayTexture, vUv);
-      vec4 nightColor = texture2D(nightTexture, vUv)* 0.2;
-      gl_FragColor = mix(nightColor, dayColor, intensity);
-    }
-  `
-});
-
-
-// ******  MOONS  ******
-// Earth
-const earthMoon = [{
-  size: 1.6,
-  texture: earthMoonTexture,
-  bump: earthMoonBump,
-  orbitSpeed: 0.001 * settings.accelerationOrbit,
-  orbitRadius: 10
-}]
-
-// Mars' moons with path to 3D models (phobos & deimos)
-const marsMoons = [
-  {
-    modelPath: '/images/mars/phobos.glb',
-    scale: 0.1,
-    orbitRadius: 5,
-    orbitSpeed: 0.002 * settings.accelerationOrbit,
-    position: 100,
-    mesh: null
-  },
-  {
-    modelPath: '/images/mars/deimos.glb',
-    scale: 0.1,
-    orbitRadius: 9,
-    orbitSpeed: 0.0005 * settings.accelerationOrbit,
-    position: 120,
-    mesh: null
-  }
-];
-
-// Jupiter
-const jupiterMoons = [
-  {
-    size: 1.6,
-    texture: ioTexture,
-    orbitRadius: 20,
-    orbitSpeed: 0.0005 * settings.accelerationOrbit
-  },
-  {
-    size: 1.4,
-    texture: europaTexture,
-    orbitRadius: 24,
-    orbitSpeed: 0.00025 * settings.accelerationOrbit
-  },
-  {
-    size: 2,
-    texture: ganymedeTexture,
-    orbitRadius: 28,
-    orbitSpeed: 0.000125 * settings.accelerationOrbit
-  },
-  {
-    size: 1.7,
-    texture: callistoTexture,
-    orbitRadius: 32,
-    orbitSpeed: 0.00006 * settings.accelerationOrbit
-  }
-];
-
-// ******  PLANET CREATIONS  ******
-const mercury = new createPlanet('Mercury', 2.4, 40, 0, mercuryTexture, mercuryBump);
-const venus = new createPlanet('Venus', 6.1, 65, 3, venusTexture, venusBump, null, venusAtmosphere);
-const earth = new createPlanet('Earth', 6.4, 90, 23, earthMaterial, null, null, earthAtmosphere, earthMoon);
-const mars = new createPlanet('Mars', 3.4, 115, 25, marsTexture, marsBump)
-// Load Mars moons
-marsMoons.forEach(moon => {
-  loadObject(moon.modelPath, moon.position, moon.scale, function(loadedModel) {
-    moon.mesh = loadedModel;
-    mars.planetSystem.add(moon.mesh);
-    moon.mesh.traverse(function (child) {
-      if (child.isMesh) {
-        child.castShadow = true;
-        child.receiveShadow = true;
-      }
-    });
-  });
-});
-
-const jupiter = new createPlanet('Jupiter', 69/4, 200, 3, jupiterTexture, null, null, null, jupiterMoons);
-const saturn = new createPlanet('Saturn', 58/4, 270, 26, saturnTexture, null, {
-  innerRadius: 18, 
-  outerRadius: 29, 
-  texture: satRingTexture
-});
-const uranus = new createPlanet('Uranus', 25/4, 320, 82, uranusTexture, null, {
-  innerRadius: 6, 
-  outerRadius: 8, 
-  texture: uraRingTexture
-});
-const neptune = new createPlanet('Neptune', 24/4, 340, 28, neptuneTexture);
-const pluto = new createPlanet('Pluto', 1, 350, 57, plutoTexture)
-
-updateEarthDefaultView(true);
-
-  // ******  PLANETS DATA  ******
-  const planetData = {
-    'Mercury': {
-        radius: '2,439.7 km',
-        tilt: '0.034°',
-        rotation: '58.6 Earth days',
-        orbit: '88 Earth days',
-        distance: '57.9 million km',
-        moons: '0',
-        info: 'The smallest planet in our solar system and nearest to the Sun.'
-    },
-    'Venus': {
-        radius: '6,051.8 km',
-        tilt: '177.4°',
-        rotation: '243 Earth days',
-        orbit: '225 Earth days',
-        distance: '108.2 million km',
-        moons: '0',
-        info: 'Second planet from the Sun, known for its extreme temperatures and thick atmosphere.'
-    },
-    'Earth': {
-        radius: '6,371 km',
-        tilt: '23.5°',
-        rotation: '24 hours',
-        orbit: '365 days',
-        distance: '150 million km',
-        moons: '1 (Moon)',
-        info: 'Third planet from the Sun and the only known planet to harbor life.'
-    },
-    'Mars': {
-        radius: '3,389.5 km',
-        tilt: '25.19°',
-        rotation: '1.03 Earth days',
-        orbit: '687 Earth days',
-        distance: '227.9 million km',
-        moons: '2 (Phobos and Deimos)',
-        info: 'Known as the Red Planet, famous for its reddish appearance and potential for human colonization.'
-    },
-    'Jupiter': {
-        radius: '69,911 km',
-        tilt: '3.13°',
-        rotation: '9.9 hours',
-        orbit: '12 Earth years',
-        distance: '778.5 million km',
-        moons: '95 known moons (Ganymede, Callisto, Europa, Io are the 4 largest)',
-        info: 'The largest planet in our solar system, known for its Great Red Spot.'
-    },
-    'Saturn': {
-        radius: '58,232 km',
-        tilt: '26.73°',
-        rotation: '10.7 hours',
-        orbit: '29.5 Earth years',
-        distance: '1.4 billion km',
-        moons: '146 known moons',
-        info: 'Distinguished by its extensive ring system, the second-largest planet in our solar system.'
-    },
-    'Uranus': {
-        radius: '25,362 km',
-        tilt: '97.77°',
-        rotation: '17.2 hours',
-        orbit: '84 Earth years',
-        distance: '2.9 billion km',
-        moons: '27 known moons',
-        info: 'Known for its unique sideways rotation and pale blue color.'
-    },
-    'Neptune': {
-        radius: '24,622 km',
-        tilt: '28.32°',
-        rotation: '16.1 hours',
-        orbit: '165 Earth years',
-        distance: '4.5 billion km',
-        moons: '14 known moons',
-        info: 'The most distant planet from the Sun in our solar system, known for its deep blue color.'
-    },
-    'Pluto': {
-        radius: '1,188.3 km',
-        tilt: '122.53°',
-        rotation: '6.4 Earth days',
-        orbit: '248 Earth years',
-        distance: '5.9 billion km',
-        moons: '5 (Charon, Styx, Nix, Kerberos, Hydra)',
-        info: 'Originally classified as the ninth planet, Pluto is now considered a dwarf planet.'
-    }
-};
-
-
-// Array of planets and atmospheres for raycasting
-// ******  SHADOWS  ******
-renderer.shadowMap.enabled = true;
-pointLight.castShadow = true;
-
-//properties for the point light
-pointLight.shadow.mapSize.width = 1024;
-pointLight.shadow.mapSize.height = 1024;
-pointLight.shadow.camera.near = 10;
-pointLight.shadow.camera.far = 20;
-
-//casting and receiving shadows
-earth.planet.castShadow = true;
-earth.planet.receiveShadow = true;
-earth.Atmosphere.castShadow = true;
-earth.Atmosphere.receiveShadow = true;
-earth.moons.forEach(moon => {
-moon.mesh.castShadow = true;
-moon.mesh.receiveShadow = true;
-});
-mercury.planet.castShadow = true;
-mercury.planet.receiveShadow = true;
-venus.planet.castShadow = true;
-venus.planet.receiveShadow = true;
-venus.Atmosphere.receiveShadow = true;
-mars.planet.castShadow = true;
-mars.planet.receiveShadow = true;
-jupiter.planet.castShadow = true;
-jupiter.planet.receiveShadow = true;
-jupiter.moons.forEach(moon => {
-  moon.mesh.castShadow = true;
-  moon.mesh.receiveShadow = true;
-  });
-saturn.planet.castShadow = true;
-saturn.planet.receiveShadow = true;
-saturn.Ring.receiveShadow = true;
-uranus.planet.receiveShadow = true;
-neptune.planet.receiveShadow = true;
-pluto.planet.receiveShadow = true;
-
-
-
-
-function animate(){
-
-  //rotating planets around the sun and itself
-  sun.rotateY(0.001 * settings.acceleration);
-  mercury.planet.rotateY(0.001 * settings.acceleration);
-  mercury.planet3d.rotateY(0.004 * settings.accelerationOrbit);
-  venus.planet.rotateY(0.0005 * settings.acceleration)
-  venus.Atmosphere.rotateY(0.0005 * settings.acceleration);
-  venus.planet3d.rotateY(0.0006 * settings.accelerationOrbit);
-  earth.planet.rotateY(0.005 * settings.acceleration);
-  earth.Atmosphere.rotateY(0.001 * settings.acceleration);
-  earth.planet3d.rotateY(0.001 * settings.accelerationOrbit);
-  mars.planet.rotateY(0.01 * settings.acceleration);
-  mars.planet3d.rotateY(0.0007 * settings.accelerationOrbit);
-  jupiter.planet.rotateY(0.005 * settings.acceleration);
-  jupiter.planet3d.rotateY(0.0003 * settings.accelerationOrbit);
-  saturn.planet.rotateY(0.01 * settings.acceleration);
-  saturn.planet3d.rotateY(0.0002 * settings.accelerationOrbit);
-  uranus.planet.rotateY(0.005 * settings.acceleration);
-  uranus.planet3d.rotateY(0.0001 * settings.accelerationOrbit);
-  neptune.planet.rotateY(0.005 * settings.acceleration);
-  neptune.planet3d.rotateY(0.00008 * settings.accelerationOrbit);
-  pluto.planet.rotateY(0.001 * settings.acceleration)
-  pluto.planet3d.rotateY(0.00006 * settings.accelerationOrbit)
-
-// Animate Earth's moon
-if (earth.moons) {
-  earth.moons.forEach(moon => {
-    const time = performance.now();
-    const tiltAngle = 5 * Math.PI / 180;
-
-    const moonX = earth.planet.position.x + moon.orbitRadius * Math.cos(time * moon.orbitSpeed);
-    const moonY = moon.orbitRadius * Math.sin(time * moon.orbitSpeed) * Math.sin(tiltAngle);
-    const moonZ = earth.planet.position.z + moon.orbitRadius * Math.sin(time * moon.orbitSpeed) * Math.cos(tiltAngle);
-
-    moon.mesh.position.set(moonX, moonY, moonZ);
-    moon.mesh.rotateY(0.01);
-  });
-}
-// Animate Mars' moons
-if (marsMoons){
-marsMoons.forEach(moon => {
-  if (moon.mesh) {
-    const time = performance.now();
-
-    const moonX = mars.planet.position.x + moon.orbitRadius * Math.cos(time * moon.orbitSpeed);
-    const moonY = moon.orbitRadius * Math.sin(time * moon.orbitSpeed);
-    const moonZ = mars.planet.position.z + moon.orbitRadius * Math.sin(time * moon.orbitSpeed);
-
-    moon.mesh.position.set(moonX, moonY, moonZ);
-    moon.mesh.rotateY(0.001);
-  }
-});
-}
-
-// Animate Jupiter's moons
-if (jupiter.moons) {
-  jupiter.moons.forEach(moon => {
-    const time = performance.now();
-    const moonX = jupiter.planet.position.x + moon.orbitRadius * Math.cos(time * moon.orbitSpeed);
-    const moonY = moon.orbitRadius * Math.sin(time * moon.orbitSpeed);
-    const moonZ = jupiter.planet.position.z + moon.orbitRadius * Math.sin(time * moon.orbitSpeed);
-
-    moon.mesh.position.set(moonX, moonY, moonZ);
-    moon.mesh.rotateY(0.01);
-  });
-}
-
-updateEarthDefaultView();
-
-// Update asteroids and overlays
-updateAsteroids();
-
-if (selectedAsteroidEntry && selectedAsteroidEntry.mesh) {
-  selectedAsteroidEntry.mesh.getWorldPosition(asteroidFocusPoint);
-  controls.target.lerp(asteroidFocusPoint, 0.15);
-
-  if (!isMovingTowardsAsteroid) {
-    asteroidCameraTarget.copy(asteroidFocusPoint).add(asteroidDefaultCameraOffset);
-    camera.position.lerp(asteroidCameraTarget, 0.02);
-  }
-} else if (!selectedPlanet && !isMovingTowardsPlanet && !isMovingTowardsAsteroid && !isZoomingOut) {
-  controls.target.lerp(earthDefaultTargetPosition, 0.1);
-  camera.position.lerp(earthDefaultCameraPosition, 0.02);
-}
-
-// ****** ASTEROID OUTLINE ******
-if (selectedAsteroidEntry && selectedAsteroidEntry.mesh) {
-  outlinePass.selectedObjects = [selectedAsteroidEntry.mesh];
-} else {
-  outlinePass.selectedObjects = [];
-}
-// ******  ZOOM IN/OUT  ******
-if (isMovingTowardsPlanet) {
-  // Smoothly move the camera towards the target position
-  camera.position.lerp(targetCameraPosition, 0.03);
-
-  // Check if the camera is close to the target position
-  if (camera.position.distanceTo(targetCameraPosition) < 1) {
-      isMovingTowardsPlanet = false;
-      showPlanetInfo(selectedPlanet.name);
-
-  }
-} else if (isMovingTowardsAsteroid) {
-  camera.position.lerp(asteroidCameraTarget, 0.03);
-
-  if (camera.position.distanceTo(asteroidCameraTarget) < 1) {
-      isMovingTowardsAsteroid = false;
-  }
-} else if (isZoomingOut) {
-  controls.target.lerp(earthDefaultTargetPosition, 0.1);
-  camera.position.lerp(zoomOutTargetPosition, 0.05);
-
-  if (camera.position.distanceTo(zoomOutTargetPosition) < 1) {
-      isZoomingOut = false;
-  }
-}
-
-  controls.update();
-  requestAnimationFrame(animate);
-  composer.render();
-}
-animate();
-
-window.addEventListener('mousemove', onMouseMove, false);
-window.addEventListener('mousedown', onDocumentMouseDown, false);
-window.addEventListener('resize', function(){
-  camera.aspect = window.innerWidth/window.innerHeight;
-  camera.updateProjectionMatrix();
-  renderer.setSize(window.innerWidth,window.innerHeight);
-  composer.setSize(window.innerWidth,window.innerHeight);
-});
